Sites: Use siteObj as source of truth to fix homepage flicker
When you set a new page as the front page on your site from /pages, we clear
out the sites in local storage and refetch the list of sites on your account. If
you have a lot of sites, it can take some time. This delay in returning the new
site list can result in a flicker where the homepage icon temporarily flips to
another, stale page.

This establishes the updated siteObj as the source of truth if the page_on_front
settings don't match. This allows us to show the correct front page while the
sites list is updating in the background.